refactor(cfd-1d): Decompose solver/core/mod.rs into leaf modules
Part of ATLAS-ARCH-007.
